数字化工具/语音转文本/server/requirements.txt
fastapi
uvicorn
python-multipart
openai
pydub
python-dotenv
faster-whisper
